Move AdLib Gold postprocessing into its own file
