+++ src/hardware/adlib.cpp
@@ -23,15 +23,14 @@
 #include <math.h>
 #include <sys/types.h>
 
+#include "adlib_gold.h"
 #include "cpu.h"
-#include "setup.h"
-#include "support.h"
+#include "dbopl.h"
 #include "mapper.h"
 #include "mem.h"
-#include "dbopl.h"
+#include "setup.h"
+#include "support.h"
 #include "../libs/nuked/opl3.h"
-#include "../libs/TDA8425_emu/TDA8425_emu.h"
-#include "../libs/YM7128B_emu/YM7128B_emu.h"
 
 #include "mame/emu.h"
 #include "mame/fmopl.h"
@@ -42,219 +41,6 @@
 #define OPL3_INTERNAL_FREQ    14400000  // The OPL3 operates at 14.4MHz
 
 static Adlib::Module *module = nullptr;
-
-class AdlibGoldSurroundProcessor {
-public:
-	AdlibGoldSurroundProcessor(const uint16_t sample_rate);
-	~AdlibGoldSurroundProcessor();
-
-	void ControlWrite(const uint8_t val) noexcept;
-	AudioFrame Process(const AudioFrame &frame) noexcept;
-
-private:
-	YM7128B_ChipIdeal *chip = nullptr;
-
-	struct {
-		uint8_t sci  = 0;
-		uint8_t a0   = 0;
-		uint8_t addr = 0;
-		uint8_t data = 0;
-	} control_state = {};
-};
-
-AdlibGoldSurroundProcessor::AdlibGoldSurroundProcessor(const uint16_t sample_rate)
-{
-	chip = (YM7128B_ChipIdeal *)malloc(sizeof(YM7128B_ChipIdeal));
-
-	YM7128B_ChipIdeal_Ctor(chip);
-	YM7128B_ChipIdeal_Setup(chip, sample_rate);
-	YM7128B_ChipIdeal_Reset(chip);
-	YM7128B_ChipIdeal_Start(chip);
-}
-
-AdlibGoldSurroundProcessor::~AdlibGoldSurroundProcessor()
-{
-	if (chip) {
-		YM7128B_ChipIdeal_Stop(chip);
-		YM7128B_ChipIdeal_Dtor(chip);
-		free(chip);
-		chip = nullptr;
-	}
-}
-
-void AdlibGoldSurroundProcessor::ControlWrite(const uint8_t val) noexcept
-{
-	// Serial data
-	const auto din = val & 1;
-	// Bit clock
-	const auto sci = val & 2;
-	// Word clock
-	const auto a0 = val & 4;
-
-	// Change register data at the falling edge of 'a0' word clock
-	if (control_state.a0 && !a0) {
-//		DEBUG_LOG_MSG("ADLIBGOLD.SURROUND: Write control register %d, data: %d",
-//		              control_state.addr,
-//		              control_state.data);
-		YM7128B_ChipIdeal_Write(chip, control_state.addr, control_state.data);
-	} else {
-		// Data is sent in serially through 'din' in MSB->LSB order,
-		// synchronised by the 'sci' bit clock. Data should be read on
-		// the rising edge of 'sci'.
-		if (!control_state.sci && sci) {
-			// The 'a0' word clock determines the type of the data.
-			if (a0)
-				// Data cycle
-				control_state.data = (control_state.data << 1) | din;
-			else
-				// Address cycle
-				control_state.addr = (control_state.addr << 1) | din;
-		}
-	}
-
-	control_state.sci = sci;
-	control_state.a0  = a0;
-}
-
-AudioFrame AdlibGoldSurroundProcessor::Process(const AudioFrame &frame) noexcept
-{
-	YM7128B_ChipIdeal_Process_Data data = {};
-	data.inputs[0]                      = frame.left + frame.right;
-
-	YM7128B_ChipIdeal_Process(chip, &data);
-
-	return {data.outputs[0], data.outputs[1]};
-}
-
-class AdlibGoldStereoProcessor {
-public:
-	AdlibGoldStereoProcessor(const uint16_t sample_rate);
-	~AdlibGoldStereoProcessor();
-
-	void Reset() noexcept;
-	void ControlWrite(const TDA8425_Reg reg, const TDA8425_Register data) noexcept;
-	AudioFrame Process(const AudioFrame &frame) noexcept;
-
-private:
-	TDA8425_Chip *chip = nullptr;
-};
-
-AdlibGoldStereoProcessor::AdlibGoldStereoProcessor(const uint16_t sample_rate)
-{
-	chip = (TDA8425_Chip *)malloc(sizeof(TDA8425_Chip));
-
-	TDA8425_Chip_Ctor(chip);
-	TDA8425_Chip_Setup(chip,
-	                   sample_rate,
-	                   TDA8425_Pseudo_C1_Table[TDA8425_Pseudo_Preset_1],
-	                   TDA8425_Pseudo_C2_Table[TDA8425_Pseudo_Preset_1],
-	                   TDA8425_Tfilter_Mode_Disabled);
-	TDA8425_Chip_Reset(chip);
-	TDA8425_Chip_Start(chip);
-
-	Reset();
-}
-
-AdlibGoldStereoProcessor::~AdlibGoldStereoProcessor()
-{
-	if (chip) {
-		TDA8425_Chip_Stop(chip);
-		TDA8425_Chip_Dtor(chip);
-		free(chip);
-		chip = nullptr;
-	}
-}
-
-void AdlibGoldStereoProcessor::Reset() noexcept
-{
-	constexpr auto volume_0db    = 60;
-	constexpr auto bass_0db      = 6;
-	constexpr auto treble_0db    = 6;
-	constexpr auto stereo_output = TDA8425_Selector_Stereo_1;
-	constexpr auto linear_stereo = TDA8425_Mode_LinearStereo
-	                            << TDA8425_Reg_SF_STL;
-
-	ControlWrite(TDA8425_Reg_VL, volume_0db);
-	ControlWrite(TDA8425_Reg_VR, volume_0db);
-	ControlWrite(TDA8425_Reg_BA, bass_0db);
-	ControlWrite(TDA8425_Reg_TR, treble_0db);
-	ControlWrite(TDA8425_Reg_SF, stereo_output & linear_stereo);
-}
-
-void AdlibGoldStereoProcessor::ControlWrite(const TDA8425_Reg addr,
-                                            const TDA8425_Register data) noexcept
-{
-	TDA8425_Chip_Write(chip, addr, data);
-}
-
-AudioFrame AdlibGoldStereoProcessor::Process(const AudioFrame &frame) noexcept
-{
-	TDA8425_Chip_Process_Data data = {};
-	data.inputs[0][0]              = frame.left;
-	data.inputs[1][0]              = frame.left;
-	data.inputs[0][1]              = frame.right;
-	data.inputs[1][1]              = frame.right;
-
-	TDA8425_Chip_Process(chip, &data);
-
-	return {data.outputs[0], data.outputs[1]};
-}
-
-class AdlibGold {
-public:
-	AdlibGold(const uint16_t sample_rate);
-
-	void SurroundControlWrite(const uint8_t val) noexcept;
-	void StereoControlWrite(const TDA8425_Reg reg,
-	                        const TDA8425_Register data) noexcept;
-
-	void Process(const int16_t *in, const uint32_t frames, float *out) noexcept;
-
-private:
-	std::unique_ptr<AdlibGoldSurroundProcessor> surround_processor;
-	std::unique_ptr<AdlibGoldStereoProcessor> stereo_processor;
-};
-
-AdlibGold::AdlibGold(const uint16_t sample_rate)
-{
-	surround_processor = std::make_unique<AdlibGoldSurroundProcessor>(sample_rate);
-	stereo_processor = std::make_unique<AdlibGoldStereoProcessor>(sample_rate);
-}
-
-void AdlibGold::StereoControlWrite(const TDA8425_Reg reg,
-                                   const TDA8425_Register data) noexcept
-{
-	stereo_processor->ControlWrite(reg, data);
-}
-
-void AdlibGold::SurroundControlWrite(const uint8_t val) noexcept
-{
-	surround_processor->ControlWrite(val);
-}
-
-void AdlibGold::Process(const int16_t *in, const uint32_t frames, float *out) noexcept
-{
-	auto frames_left = frames;
-	while (frames_left--) {
-		AudioFrame frame = {static_cast<float>(in[0]),
-		                    static_cast<float>(in[1])};
-
-		const auto wet = surround_processor->Process(frame);
-		// Additional wet signal level boost to make the emulated sound
-		// more closely resemble real hardware recordings.
-		constexpr auto wet_boost = 1.6;
-		frame.left += wet.left * wet_boost;
-		frame.right += wet.right * wet_boost;
-
-		frame = stereo_processor->Process(frame);
-
-		out[0] = frame.left;
-		out[1] = frame.right;
-		in += 2;
-		out += 2;
-	}
-}
-
 static AdlibGold *adlib_gold = nullptr;
 
 constexpr auto render_frames = 1024;
@@ -1234,3 +1020,4 @@ void OPL_ShutDown(Section* /*sec*/){
 	module = 0;
 
 }
+
